52 autorization

+++ app/Http/Controllers/ComplaintController.php
@@ -10,9 +10,11 @@ use Illuminate\Http\Request;
 class ComplaintController extends Controller
 {
     public function index(){
+        $this->authorize('view',Complaint::class);
         return new ComplaintCollection(Complaint::paginate(10));
     }
     public function show(Complaint $complaint){
+        $this->authorize('view',$complaint);
         return response()->json(new ComplaintResource($complaint),200);
     }
 
@@ -36,19 +38,18 @@ class ComplaintController extends Controller
         $messages= [
             'required'=> 'El campo :attribute es obligatorio.',
         ];
-
         $request->validate([
             'complaint' =>'required',
             'username' =>'required|string|max:35',
             'email'=>'required|string|max:35',
         ],$messages);
-
         $complaint = Complaint::create($request->all());
         return response()->json($complaint, 201);
     }
 
     public function update(Request $request, Complaint $complaint)
     {
+        $this->authorize('update',$complaint);
         $messages= [
             'required'=> 'El campo :attribute es obligatorio.',
         ];
@@ -63,6 +64,7 @@ class ComplaintController extends Controller
         return response()->json($complaint, 200);
     }
     public function delete(Complaint $complaint){
+        $this->authorize('delete',$complaint);
         $complaint->delete();
         return response()->json(null, 204);
     }
